Code cleanup

+++ src/Weaviate.Client/Models/Collection.Update.cs
@@ -20,13 +20,13 @@ public record Bm25ConfigUpdate(BM25Config WrappedBm25)
     public double? B
     {
         get => WrappedBm25.B;
-        set => WrappedBm25.B = (float)(value ?? BM25Config.Default.B);
+        set => WrappedBm25.B = value ?? BM25Config.Default.B;
     }
 
     public double? K1
     {
         get => WrappedBm25.K1;
-        set => WrappedBm25.K1 = (float)(value ?? BM25Config.Default.K1);
+        set => WrappedBm25.K1 = value ?? BM25Config.Default.K1;
     }
 }
